refactor: share ExpectLaunched() between VM and machine instances
